+++ popcorn/readers.py
@@ -1,4 +1,4 @@
-from json import load as load_json
+from ijson import items as iterate_json_items
 
 from popcorn.structures import Event, Reader
 
@@ -7,34 +7,27 @@ def _getv(item: dict, prop: str, default: str | int | bool = -1) -> str | int | 
     return item[prop] if (prop in item.keys()) else default
 
 
-class LevelZeroTracerJsonReader(Reader):
+class UnitraceJsonReader(Reader):
     def __init__(self):
         super().__init__(format="json")
 
     def create_event_from_trace_item(self, item) -> Event:
         event = Event()
+        event.dur = _getv(item, "dur", default=0)
+        event.num_calls = 1
         event.ph = _getv(item, "ph", default="N/A")
-        event.tid = _getv(item, "tid")
         event.pid = _getv(item, "pid")
         event.name = _getv(item, "name", default="N/A")
         event.cat = _getv(item, "cat", default="N/A")
         event.ts = _getv(item, "ts")
-        event.id = _getv(item, "id")
-        event.dur = _getv(item, "dur", default=0)
-        event.args_id = (
-            item["args"]["id"]
-            if (("args" in item.keys()) and ("id" in item["args"].keys()))
-            else -1
-        )
         return event
 
     def read(self, filename: str, uniques: bool = True, cat: str | None = None) -> list[Event]:
         if uniques:
             unique_events: dict[str, Event] = {}
 
-            with open(filename, "r") as f:
-                data = load_json(f)
-                for item in data["traceEvents"]:
+            with open(filename, "rb") as f:
+                for item in iterate_json_items(f, "traceEvents.item"):
                     item_name = _getv(item, "name", default="N/A")
                     item_category = _getv(item, "cat", default=False)
                     same_category = item_category and (item_category == cat)
@@ -42,6 +35,7 @@ class LevelZeroTracerJsonReader(Reader):
                         not cat
                     ) or same_category:  # no filter applied or category matches
                         if item_name in unique_events:  # collapse uniques duration
+                            unique_events[item_name].num_calls += 1
                             unique_events[item_name].dur += _getv(
                                 item, "dur", default=0
                             )
@@ -54,9 +48,8 @@ class LevelZeroTracerJsonReader(Reader):
         else:
             trace_events: list[Event] = []
 
-            with open(filename, "r") as f:
-                data = load_json(f)
-                for item in data["traceEvents"]:
+            with open(filename, "rb") as f:
+                for item in iterate_json_items(f, "traceEvents.item"):
                     item_category = _getv(item, "cat", default=False)
                     same_category = item_category and (item_category == cat)
                     if (