+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Any notable change to this project will be documented in this file.
 
 ## Unreleased
 
+## [v0.0.3 Release (13-02-2024)](https://github.com/efleming-intel/popcorn/tree/v0.0.3)
+
+- updated README.md
+- added coveralls.io support
+- rehauled console output for a cleaner look
+- replaced json usage with ijson for more stable json file reading and allow reading of large json files
+
 ## [v0.0.2 Release (27-10-2023)](https://github.com/efleming-intel/popcorn/tree/v0.0.2)
 
 - print display item count in console output